Move express stuff from server.js to services/express
