+++ src/server.js
@@ -4,36 +4,13 @@ import { Provider } from 'react-redux'
 import { createMemoryHistory, RouterContext, match } from 'react-router'
 import { syncHistoryWithStore } from 'react-router-redux'
 import serialize from 'serialize-javascript'
-import express from 'express'
-import forceSSL from 'express-force-ssl'
-import cookieParser from 'cookie-parser'
-import bodyParser from 'body-parser'
-import csrf from 'csurf'
-import path from 'path'
-import compression from 'compression'
 import styleSheet from 'styled-components/lib/models/StyleSheet'
-import { env, ip, port, root } from 'config'
+import { env } from 'config'
+import app from 'services/express'
 import routes from 'routes'
 import configureStore from 'store/configure'
 import { setCsrfToken } from 'store'
 import { Html } from 'components'
-
-const app = express()
-
-if (env === 'production') {
-  app.set('forceSSLOptions', {
-    enable301Redirects: true,
-    trustXFPHeader: true
-  })
-  app.use(forceSSL)
-}
-
-app.use(cookieParser())
-app.use(bodyParser.urlencoded({ extended: false }))
-app.use(bodyParser.json())
-app.use(csrf({ cookie: true }))
-app.use(compression())
-app.use(express.static(path.join(root, 'dist')))
 
 app.use((req, res, next) => {
   if (env === 'development') {
@@ -98,12 +75,4 @@ app.use((req, res, next) => {
       res.status(500).end()
     })
   })
-})
-
-app.listen(port, (error) => {
-  if (error) {
-    console.error(error)
-  } else {
-    console.info(`Listening on http://${ip}:${port}`)
-  }
 })
